djangoapi/requirements.txt
django
djangorestframework
psycopg2
psycogp2-binary
django-cors-headers